backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
pathspec==0.11.2 
django-filter>=25.1